requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8,<3.0
djangorestframework>=3.12,<4.0
pandas>=1.1,<2.0
django-environ>=0.4,<1.0
